examen3/pregunta2/requirements.txt
opencv-python
numpy
Pillow
mysql-connector-python
